+++ lib/rapa/responses/base_response.rb
@@ -1,6 +1,8 @@
 module Rapa
   module Responses
     class BaseResponse
+      include ::Enumerable
+
       # @param faraday_response [Faraday::Response]
       def initialize(faraday_response)
         @faraday_response = faraday_response
@@ -9,6 +11,11 @@ module Rapa
       # @return [String]
       def body
         faraday_response.body
+      end
+
+      # @note Implementation for Enumerable
+      def each(&block)
+        resources.each(&block)
       end
 
       # @todo
@@ -22,6 +29,13 @@ module Rapa
         faraday_response.headers
       end
 
+      # @return [Array<Rapa::Resources::ItemResource>]
+      def resources
+        sources.map do |source|
+          resource_class.new(source)
+        end
+      end
+
       # @return [Integer]
       def status
         faraday_response.status
@@ -29,9 +43,33 @@ module Rapa
 
       private
 
+      # @private
       # @return [Faraday::Response]
       def faraday_response
         @faraday_response
+      end
+
+      # @return [Class]
+      def resource_class
+        raise ::NotImplementedError
+      end
+
+      # @private
+      # @return [Array<Hash>, Hash]
+      def source_or_sources
+        body["ItemLookupResponse"]["Items"]["Item"]
+      end
+
+      # @return [Array<Hash>]
+      def sources
+        case source_or_sources
+        when ::Array
+          source_or_sources
+        when nil
+          []
+        else
+          [source_or_sources]
+        end
       end
     end
   end