+++ lib/rapa/responses/list_items_response.rb
@@ -1,6 +1,12 @@
 module Rapa
   module Responses
     class ListItemsResponse < BaseResponse
+      private
+
+      # @note Override
+      def resource_class
+        ::Rapa::Resources::ItemResource
+      end
     end
   end
 end