+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Add `Rapa::Resources::ItemResource`
+
 ## 0.0.1
 
 - 1st Release :tada: